app/requirements.txt
blinker==1.7.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.2
Flask==3.0.0
Flask-Cors==4.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.22
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
zipp==3.17.0
